OY-4290_OY-4319 käytetään kkhakijat-rajapinnassa masterhenkilön tietoja, oppijanumerorekisteri käyttöön KkHakijaServicessa
