+++ src/test/scala/fi/vm/sade/hakurekisteri/rest/KkHakijaResourceSpec.scala
@@ -83,6 +83,7 @@ class KkHakijaResourceSpec
     hakukohdeAggregatorMock,
     hakuMock,
     koodistoMock,
+    MockOppijaNumeroRekisteri,
     suoritusMock,
     valintaTulosMock,
     valintaRekisteri,
