+++ src/test/scala/fi/vm/sade/hakurekisteri/web/kkhakija/KkHakijaServiceSpec.scala
@@ -195,6 +195,7 @@ class KkHakijaServiceSpec
     hakukohdeAggregatorMock,
     hakuMock,
     koodistoMock,
+    MockOppijaNumeroRekisteri,
     suoritusMock,
     valintaTulosMock,
     valintaRekisteri,
@@ -251,6 +252,7 @@ class KkHakijaServiceSpec
       hakukohdeAggregatorMock,
       hakuMock,
       koodistoMock,
+      MockOppijaNumeroRekisteri,
       suoritusMock,
       valintaTulosMock,
       valintaRekisteri,
@@ -765,6 +767,7 @@ class KkHakijaServiceSpec
       hakukohdeAggregatorMock,
       hakuMock,
       koodistoMock,
+      MockOppijaNumeroRekisteri,
       suoritusMock,
       valintaTulosMock,
       valintaRekisteri,
